+++ mobile/src/tabs/connect/index.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { ScrollView } from 'react-native';
 import { createStackNavigator } from 'react-navigation';
-import PropTypes from 'prop-types';
+import BackgroundView from 'ui/BackgroundView';
 import TableView, { Cell, CellIcon, CellText, Divider } from 'ui/TableView';
-import LiveNowButton from '../../live';
+import tabBarIcon from '../tabBarIcon';
 
 export class ConnectScreen extends React.Component {
   static navigationOptions = {
@@ -11,61 +11,51 @@ export class ConnectScreen extends React.Component {
   };
   render() {
     return (
-      <ScrollView>
-        <LiveNowButton navigation={this.props.navigation} />
-        <TableView>
-          <Cell>
-            <CellIcon name="check" />
-            <CellText>Find a serving opportunity</CellText>
-          </Cell>
-          <Divider />
-          <Cell>
-            <CellIcon name="groups" />
-            <CellText>Join a small group</CellText>
-          </Cell>
-          <Divider />
-          <Cell>
-            <CellIcon name="share" />
-            <CellText>I need prayer</CellText>
-          </Cell>
-          <Divider />
-          <Cell>
-            <CellIcon name="download" />
-            <CellText>I would like to give</CellText>
-          </Cell>
-          <Divider />
-          <Cell>
-            <CellIcon name="building" />
-            <CellText>Find Service Times & Locations</CellText>
-          </Cell>
-        </TableView>
-      </ScrollView>
+      <BackgroundView>
+        <ScrollView>
+          <TableView>
+            <Cell>
+              <CellIcon name="check" />
+              <CellText>Find a serving opportunity</CellText>
+            </Cell>
+            <Divider />
+            <Cell>
+              <CellIcon name="groups" />
+              <CellText>Join a small group</CellText>
+            </Cell>
+            <Divider />
+            <Cell>
+              <CellIcon name="share" />
+              <CellText>I need prayer</CellText>
+            </Cell>
+            <Divider />
+            <Cell>
+              <CellIcon name="download" />
+              <CellText>I would like to give</CellText>
+            </Cell>
+            <Divider />
+            <Cell>
+              <CellIcon name="building" />
+              <CellText>Find Service Times & Locations</CellText>
+            </Cell>
+          </TableView>
+        </ScrollView>
+      </BackgroundView>
     );
   }
 }
 
-ConnectScreen.propTypes = {
-  navigation: PropTypes.shape({
-    navigate: PropTypes.func,
-  }),
-};
-
-export const ConnectStack = createStackNavigator(
+const ConnectStack = createStackNavigator(
   {
     Connect: ConnectScreen,
   },
   {
     initialRouteName: 'Connect',
-    navigationOptions: {
-      headerStyle: {
-        backgroundColor: '#f4511e',
-      },
-      headerTintColor: '#fff',
-      headerTitleStyle: {
-        fontWeight: 'bold',
-      },
-    },
   }
 );
+
+ConnectStack.navigationOptions = {
+  tabBarIcon: tabBarIcon('profile'),
+};
 
 export default ConnectStack;