+++ mobile/src/tabs/home/index.js
@@ -2,22 +2,49 @@ import React from 'react';
 import { Query } from 'react-apollo';
 import { createStackNavigator } from 'react-navigation';
 import { get } from 'lodash';
-import FeedView from 'ui/FeedView';
 import PropTypes from 'prop-types';
-import FlexedView from 'ui/FlexedView';
+
+import { withTheme } from 'ui/theme';
+import FeedView from 'ui/FeedView';
+import BackgroundView from 'ui/BackgroundView';
+
 import GET_USER_FEED from './query';
+import tabBarIcon from '../tabBarIcon';
 import LiveNowButton from '../../live';
 
-export class HomeScreen extends React.Component {
-  static navigationOptions = {
-    title: 'Home',
-  };
+// TODO: what are our thoughts around using this @-syntax for HOCs?
+@withTheme(({ theme, ...otherProps }) => ({
+  headerBackgroundColor: theme.colors.background.primary,
+  headerTintColor: theme.colors.background.paper,
+  ...otherProps,
+}))
+class HomeScreen extends React.Component {
+  static navigationOptions = ({ navigation }) => ({
+    title: 'Apollos Church',
+    headerStyle: {
+      backgroundColor: navigation.getParam('backgroundColor'),
+    },
+    headerTintColor: navigation.getParam('tintColor'),
+  });
 
   static propTypes = {
     navigation: PropTypes.shape({
+      getParam: PropTypes.func,
+      setParams: PropTypes.func,
       navigate: PropTypes.func,
     }),
+    headerBackgroundColor: PropTypes.string,
+    headerTintColor: PropTypes.string,
   };
+
+  constructor(props) {
+    super(props);
+
+    props.navigation.setParams({
+      backgroundColor: props.headerBackgroundColor,
+      tintColor: props.headerTintColor,
+    });
+  }
 
   onPress = (item) =>
     this.props.navigation.navigate('ContentSingle', {
@@ -27,19 +54,20 @@ export class HomeScreen extends React.Component {
 
   render() {
     return (
-      <FlexedView>
-        <LiveNowButton navigation={this.props.navigation} />
+      <BackgroundView>
         <Query query={GET_USER_FEED}>
-          {({ loading, error, data }) => (
+          {({ loading, error, data, refetch }) => (
             <FeedView
               content={get(data, 'userFeed.edges', [])}
               isLoading={loading}
               error={error}
+              refetch={refetch}
+              ListHeaderComponent={LiveNowButton}
               onPressItem={this.onPress}
             />
           )}
         </Query>
-      </FlexedView>
+      </BackgroundView>
     );
   }
 }
@@ -50,16 +78,11 @@ export const HomeStack = createStackNavigator(
   },
   {
     initialRouteName: 'Home',
-    navigationOptions: {
-      headerStyle: {
-        backgroundColor: '#f4511e',
-      },
-      headerTintColor: '#fff',
-      headerTitleStyle: {
-        fontWeight: 'bold',
-      },
-    },
   }
 );
+
+HomeStack.navigationOptions = {
+  tabBarIcon: tabBarIcon('home'),
+};
 
 export default HomeStack;