+++ mobile/src/tabs/index.js
@@ -1,23 +1,25 @@
 import { createBottomTabNavigator } from 'react-navigation';
-import { ConnectStack } from './connect';
-import { HomeStack } from './home';
-import { ProfileStack } from './profile';
-import { SearchStack } from './search';
-import { SectionsStack } from './sections';
+import ConnectStack from './connect';
+import HomeStack from './home';
+import SectionsStack from './sections';
 
-export { default as ConnectStack } from './connect';
-export { default as HomeStack } from './home';
-export { default as ProfileStack } from './profile';
-export { default as SearchStack } from './search';
-export { default as SectionsStack } from './sections';
-
-export const TabStack = createBottomTabNavigator({
-  Home: HomeStack,
-  Sections: SectionsStack,
-  Connect: ConnectStack,
-  Search: SearchStack,
-  Profile: ProfileStack,
-});
+export const TabStack = createBottomTabNavigator(
+  {
+    Home: HomeStack,
+    Sections: SectionsStack,
+    Connect: ConnectStack,
+  },
+  {
+    tabBarOptions: {
+      showLabel: false,
+      activeTintColor: '#17B582', // TODO: get from theme
+      inactiveTintColor: '#A5A5A5', // TODO: get from theme
+      style: {
+        backgroundColor: 'white',
+      },
+    },
+  }
+);
 
 TabStack.navigationOptions = {
   header: null,