+++ mobile/src/tabs/sections/index.js
@@ -1,46 +1,33 @@
 import React from 'react';
-import { Button, StyleSheet, Text, View } from 'react-native';
+import { Button } from 'react-native';
 import { createStackNavigator } from 'react-navigation';
 import PropTypes from 'prop-types';
-import FlexedView from 'ui/FlexedView';
+import BackgroundView from 'ui/BackgroundView';
 import Articles from 'articles';
 import Devotionals from 'devotionals';
 import News from 'news';
-import LiveNowButton from '../../live';
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
+import tabBarIcon from '../tabBarIcon';
 
 export class SectionsScreen extends React.Component {
   static navigationOptions = {
-    title: 'Sections',
+    title: 'Discover',
   };
   render() {
     return (
-      <FlexedView>
-        <LiveNowButton navigation={this.props.navigation} />
-        <View style={styles.container}>
-          <Text>Sections Screen</Text>
-          <Button
-            title="Articles"
-            onPress={() => this.props.navigation.navigate('Articles')}
-          />
-          <Button
-            title="Devotionals"
-            onPress={() => this.props.navigation.navigate('Devotionals')}
-          />
-          <Button
-            title="News"
-            onPress={() => this.props.navigation.navigate('News')}
-          />
-        </View>
-      </FlexedView>
+      <BackgroundView>
+        <Button
+          title="Articles"
+          onPress={() => this.props.navigation.navigate('Articles')}
+        />
+        <Button
+          title="Devotionals"
+          onPress={() => this.props.navigation.navigate('Devotionals')}
+        />
+        <Button
+          title="News"
+          onPress={() => this.props.navigation.navigate('News')}
+        />
+      </BackgroundView>
     );
   }
 }
@@ -60,16 +47,11 @@ export const SectionsStack = createStackNavigator(
   },
   {
     initialRouteName: 'Sections',
-    navigationOptions: {
-      headerStyle: {
-        backgroundColor: '#f4511e',
-      },
-      headerTintColor: '#fff',
-      headerTitleStyle: {
-        fontWeight: 'bold',
-      },
-    },
   }
 );
+
+SectionsStack.navigationOptions = {
+  tabBarIcon: tabBarIcon('sections'),
+};
 
 export default SectionsStack;