+++ mobile/src/tabs/tabBarIcon.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import Icon from 'ui/Icon';
+
+const tabBarIcon = (name) => {
+  function TabBarIcon({ tintColor }) {
+    return <Icon name={name} fill={tintColor} size={24} />;
+  }
+  TabBarIcon.propTypes = {
+    tintColor: PropTypes.string,
+  };
+  return TabBarIcon;
+};
+
+export default tabBarIcon;