+++ mobile/src/ui/BackgroundView/BackgroundView.tests.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import renderer from 'react-test-renderer';
+
+import Providers from 'TestProviders';
+import BackgroundView from './';
+
+describe('The BackgroundView component', () => {
+  it('should render', () => {
+    const tree = renderer.create(
+      <Providers>
+        <BackgroundView />
+      </Providers>
+    );
+    expect(tree).toMatchSnapshot();
+  });
+});