+++ CHANGELOG.md
@@ -8,4 +8,8 @@ v0.1: June 15, 2018
 * Designed initial project deployment and release processes
 * Added a version tool for quickly updating iOS and Android version numbers
 * Created a static connect menu on the connect tab.
+* Clean up tab headings to be more consistent
+* Tweak styles on live now button
+* add pull-to-refresh support on the home screen
+* add material bottom tab bar
 * Added the ability to tap on a home feed item and see it's content details